+++ CHANGELOG.md
@@ -8,6 +8,14 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 0.21.2
+
+### Changes
+
+- Adds `getPageHref` to `Pagination`
+
+## 0.21.1
+
 ### Changes
 
 - `Accordion`: Adds `fixed-height` modifier and `defaultOpen` prop
@@ -17,6 +25,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### BugFixes
 
 - Breadcrumbs now show up in light colours
+- Template reflow and breakpoints now happen at the same time
+- Fixes release github action
 
 ## 0.21.0